Limit maximum size of uploads to be viewed as code
When using "view code" the whole file is read into RAM before being
echoed back to the user. When large uploads are allowed (e.g. in the
same order of magnitude as the available RAM) this can lead to
performance problems and can easily be abused by a malicious user.

This is a breaking change - if a server operator wants to maintain the
unlimited behaviour he would have to set ViewLimit >= SizeLimit.